add breaking test for issue #130
With locale 'auto' the locale will be guessed from environment and
browser (but not default) once
Zend_Translate_Adapter::_addTranslationData() is called and the
according locale is found in the content. Later the constructor will try
to replace this found value with the given, which is none, so the
default is used anyway.
This can currently be worked around by calling the same parameters in a
two-step style.